Uncouple some testing framework from tests/user.py, and place it in the general tests/__init__.py file
