+++ backend/users.py
@@ -1,4 +1,4 @@
-from utils import *
+from .utils import *
 
 #TODO: Next step is GitHub account integration. After that, it's resume upload. After that, it's AI integration (first step, create project list in markdown. Next,rewrite it in the style of the uploaded latex. Next, pick a spot to add project list. Finally, just do replace).
